Clean up TaskDriver, move all command line support code to a new tool class TaskAdmin.java, and build a commandline shell script from TaskAdmin.
